+++ plotting/matRad_computeAllVoiSurfaces.m
@@ -8,9 +8,9 @@ function cst = matRad_computeAllVoiSurfaces(ct,cst)
     % prevent closure of waitbar and show busy state
     set(figureWait,'pointer','watch');
     
-    xCoord = ct.resolution.x * (1:ct.cubeDim(1));
-    yCoord = ct.resolution.y * (1:ct.cubeDim(2));
-    zCoord = ct.resolution.z * (1:ct.cubeDim(3));
+    xCoord = ct.resolution.x * double(1:ct.cubeDim(1));
+    yCoord = ct.resolution.y * double(1:ct.cubeDim(2));
+    zCoord = ct.resolution.z * double(1:ct.cubeDim(3));
     
     [xMesh,yMesh,zMesh] = meshgrid(xCoord,yCoord,zCoord);
